requirements.txt
python-dotenv
openai
replicate
aiohttp
aiofiles